Frontend redesign
Use "useEffect" where necessary (instead of clumsy workarounds). Improve 
anchor style consistency across different pages. Use 404 message when 
requesting a question that does not exist. Display a message telling a 
user to log in when a guest requests a specific question (but no flash 
message, due to a bug that could not be resolved). Remove unnecessary 
comments. Other minor fixes.